+++ src/AtomUI.Desktop.Controls/DatePicker/DatePickerToken.cs
@@ -79,6 +79,11 @@ internal class DatePickerToken : AbstractControlDesignToken
     public double TextHeight { get; set; }
 
     /// <summary>
+    /// 年、季度、月份面板单元格内容宽度
+    /// </summary>
+    public double YearMonthCellWidth { get; set; }
+
+    /// <summary>
     /// 面板内容内边距
     /// </summary>
     public Thickness PanelContentPadding { get; set; }
@@ -127,6 +132,7 @@ internal class DatePickerToken : AbstractControlDesignToken
         CellWidth              = SharedToken.ControlHeightSM;
         CellHeight             = SharedToken.ControlHeightSM;
         TextHeight             = SharedToken.ControlHeightLG;
+        YearMonthCellWidth     = SharedToken.ControlHeightLG * 1.5;
         WithoutTimeCellHeight  = (SharedToken.ControlHeightLG * 1.65 + SharedToken.ControlHeightSM) / 2;
         CellMargin             = SharedToken.MarginXXS;
         PanelContentPadding    = SharedToken.PaddingSM;